// contracts/amm/amm-snip20/schema/init_msg.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type HumanAddr = string;
/**
 * Binary is a wrapper around Vec<u8> to add base64 de/serialization with serde. It also adds some helper methods to help encode inline.
 *
 * This is only needed as serde-json-{core,wasm} has a horrible encoding for Vec<u8>
 */
export type Binary = string;
export type Uint128 = string;

export interface InitMsg {
  admin?: HumanAddr | null;
  callback?: Callback | null;
  config?: InitConfig | null;
  decimals: number;
  initial_allowances?: InitialAllowance[] | null;
  initial_balances?: InitialBalance[] | null;
  name: string;
  prng_seed: Binary;
  symbol: string;
}
/**
 * Info needed to have the other contract respond.
 */
export interface Callback {
  /**
   * Info about the contract requesting the callback.
   */
  contract: ContractLink;
  /**
   * The message to call.
   */
  msg: Binary;
}
/**
 * Info needed to talk to a contract instance.
 */
export interface ContractLink {
  address: HumanAddr;
  code_hash: string;
}
/**
 * This type represents optional configuration values which can be overridden. All values are optional and have defaults which are more private by default, but can be overridden if necessary
 */
export interface InitConfig {
  /**
   * Indicates whether burn functionality should be enabled default: False
   */
  enable_burn?: boolean | null;
  /**
   * Indicates whether deposit functionality should be enabled default: False
   */
  enable_deposit?: boolean | null;
  /**
   * Indicates whether mint functionality should be enabled default: False
   */
  enable_mint?: boolean | null;
  /**
   * Indicates whether redeem functionality should be enabled default: False
   */
  enable_redeem?: boolean | null;
  /**
   * Indicates whether the total supply is public or should be kept secret. default: False
   */
  public_total_supply?: boolean | null;
}
export interface InitialAllowance {
  amount: Uint128;
  expiration?: number | null;
  owner: HumanAddr;
  spender: HumanAddr;
}
export interface InitialBalance {
  address: HumanAddr;
  amount: Uint128;
}